Move GroupOfStations and MultiModalStation
