+++ src/main/java/org/opentripplanner/routing/algorithm/filterchain/filter/TransitAlertFilter.java
@@ -1,14 +1,13 @@
 package org.opentripplanner.routing.algorithm.filterchain.filter;
 
 import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
-import org.opentripplanner.model.MultiModalStation;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.Leg;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryListFilter;
 import org.opentripplanner.routing.algorithm.mapping.AlertToLegMapper;
 import org.opentripplanner.routing.services.TransitAlertService;
+import org.opentripplanner.transit.model.site.MultiModalStation;
 import org.opentripplanner.transit.model.site.Station;
 
 public class TransitAlertFilter implements ItineraryListFilter {
