+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/RoutingRequestToFilterChainMapper.java
@@ -3,7 +3,6 @@ package org.opentripplanner.routing.algorithm.mapping;
 import java.time.Instant;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import org.opentripplanner.model.MultiModalStation;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.model.plan.SortOrder;
 import org.opentripplanner.routing.algorithm.filterchain.GroupBySimilarity;
@@ -13,6 +12,7 @@ import org.opentripplanner.routing.algorithm.filterchain.ListSection;
 import org.opentripplanner.routing.api.request.ItineraryFilterParameters;
 import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.routing.services.TransitAlertService;
+import org.opentripplanner.transit.model.site.MultiModalStation;
 import org.opentripplanner.transit.model.site.Station;
 
 public class RoutingRequestToFilterChainMapper {
