+++ src/main/java/org/opentripplanner/transit/model/site/MultiModalStation.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.model;
+package org.opentripplanner.transit.model.site;
 
 import java.util.ArrayList;
 import java.util.Collection;
